refactor(login): added hook file for login page
